+++ src/js/VideoStorage.js
@@ -1,6 +1,6 @@
 import { ERROR_MESSAGE, STORAGE_MAX_COUNT } from "./constants/constants";
 
-export default class UserStorage {
+export default class VideoStorage {
   constructor() {
     this.storage = JSON.parse(localStorage.getItem("videos")) || [];
   }
@@ -9,20 +9,16 @@ export default class UserStorage {
     return this.storage.includes(responseId);
   }
 
-  addStorage(data) {
+  addVideoData(data) {
     if (this.storage.length >= STORAGE_MAX_COUNT) {
-      throw new Error(ERROR_MESSAGE.USER_STORAGE_OVERFLOW);
+      throw new Error(ERROR_MESSAGE.VIDEO_STORAGE_OVERFLOW);
     }
 
     this.storage = [...this.storage, data];
-    this.setLocalStorage();
+    localStorage.setItem("videos", JSON.stringify(this.storage));
   }
 
   getStorage() {
     return this.storage;
-  }
-
-  setLocalStorage() {
-    localStorage.setItem("videos", JSON.stringify(this.storage));
   }
 }
